ml-services/rag-service/requirements.txt
fastapi
uvicorn
sentence-transformers
pinecone
pymongo
dnspython
python-dotenv
numpy
requests
